+++ wcfsetup/install/files/lib/system/interaction/admin/PaidSubscriptionUserInteractions.class.php
@@ -7,6 +7,7 @@ use wcf\event\interaction\admin\PaidSubscriptionUserInteractionCollecting;
 use wcf\system\event\EventHandler;
 use wcf\system\interaction\AbstractInteractionProvider;
 use wcf\system\interaction\DeleteInteraction;
+use wcf\system\WCF;
 
 /**
  * Interaction provider for paid subscription users.
@@ -20,6 +21,13 @@ final class PaidSubscriptionUserInteractions extends AbstractInteractionProvider
 {
     public function __construct()
     {
+        if (
+            \MODULE_PAID_SUBSCRIPTION === 0
+            || !WCF::getSession()->getPermission('admin.paidSubscription.canManageSubscription')
+        ) {
+            return;
+        }
+
         $this->addInteractions([
             new DeleteInteraction('core/paidSubscriptions/users/%s')
         ]);
