+++ wcfsetup/install/files/lib/system/interaction/admin/SmileyInteractions.class.php
@@ -7,6 +7,7 @@ use wcf\event\interaction\admin\SmileyInteractionCollecting;
 use wcf\system\event\EventHandler;
 use wcf\system\interaction\AbstractInteractionProvider;
 use wcf\system\interaction\DeleteInteraction;
+use wcf\system\WCF;
 
 /**
  * Interaction provider for smilies.
@@ -20,6 +21,13 @@ final class SmileyInteractions extends AbstractInteractionProvider
 {
     public function __construct()
     {
+        if (
+            \MODULE_SMILEY === 0
+            || !WCF::getSession()->getPermission('admin.content.smiley.canManageSmiley')
+        ) {
+            return;
+        }
+
         $this->addInteractions([
             new DeleteInteraction("core/smilies/%s")
         ]);
